Move ViewMenu class up a package
