+++ src/games/strategy/triplea/ui/menubar/TripleAMenu.java
@@ -50,7 +50,6 @@ import games.strategy.triplea.ui.IUIContext;
 import games.strategy.triplea.ui.PoliticalStateOverview;
 import games.strategy.triplea.ui.TripleAFrame;
 import games.strategy.triplea.ui.VerifiedRandomNumbersDialog;
-import games.strategy.triplea.ui.menubar.menu.view.ViewMenu;
 import games.strategy.ui.SwingAction;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.data.GameData;
